+++ lib/unique_list.dart
@@ -1,3 +1,0 @@
-library unique_list;
-
-export 'src/unique_list_impl.dart';