+++ lib/x_unique_list.dart
@@ -1,0 +1,3 @@
+library x_unique_list;
+
+export 'src/x_unique_list.dart';